Move map logic to item-map component
