+++ src/app/components/create/item-map/item-map.component.scss
@@ -1,0 +1,7 @@
+.truncate-text {
+  p {
+    white-space: nowrap;
+    overflow: hidden;
+    text-overflow: ellipsis;
+  }
+}
